# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Ticketing API",
        "description": "API untuk mengelola berbagai resource",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/city-or-regency": {
            "get": {
                "tags": [
                    "CityOrRegency"
                ],
                "summary": "Get all cities or regencies",
                "operationId": "c575220bda9bce33e1e892b3da080b34",
                "responses": {
                    "200": {
                        "description": "List of cities or regencies"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CityOrRegency"
                ],
                "summary": "Create a new city or regency",
                "operationId": "fa35c6c4ced58928081f26bc4869ec86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "province_id",
                                    "no_city_or_regency",
                                    "city_or_regency_name"
                                ],
                                "properties": {
                                    "province_id": {
                                        "type": "integer"
                                    },
                                    "no_city_or_regency": {
                                        "type": "string"
                                    },
                                    "city_or_regency_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "City or regency created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/city-or-regency/{id}": {
            "get": {
                "tags": [
                    "CityOrRegency"
                ],
                "summary": "Get a specific city or regency",
                "operationId": "d301ff2d20e188ffce260723256f4d8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City or regency found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CityOrRegency"
                ],
                "summary": "Update a city or regency",
                "operationId": "728eec895da96eeadf65f3dd665b4afb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "province_id": {
                                        "type": "integer"
                                    },
                                    "no_city_or_regency": {
                                        "type": "string"
                                    },
                                    "city_or_regency_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "City or regency updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CityOrRegency"
                ],
                "summary": "Delete a city or regency",
                "operationId": "a2fe9efba0a69446a6bd656607dd8ab8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "City or regency deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/city-or-regency/export-format": {
            "get": {
                "tags": [
                    "CityOrRegency"
                ],
                "summary": "Export format city or regency data",
                "operationId": "5f135bb4943dcdf3e0995a6d10c93d56",
                "responses": {
                    "200": {
                        "description": "Excel format file exported"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/city-or-regency/export": {
            "get": {
                "tags": [
                    "CityOrRegency"
                ],
                "summary": "Export city or regency data",
                "operationId": "b4204ed97effa2f35e63ca3f6883da38",
                "responses": {
                    "200": {
                        "description": "Excel file exported"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/city-or-regency/import": {
            "post": {
                "tags": [
                    "CityOrRegency"
                ],
                "summary": "Import city or regency data",
                "operationId": "f1e7ceadb173f930e194861528a664f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data imported successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}